Renamed library to constrainedAxisPointingLibrary
