+++ src/fswAlgorithms/attGuidance/sepPoint/sepPoint.i
@@ -30,7 +30,7 @@ from Basilisk.architecture.swig_common_model import *
 
 %include "sys_model.h"
 %include "sepPoint.h"
-%include "fswAlgorithms/attGuidance/_GeneralModuleFiles/attitudePointingLibrary.h"
+%include "fswAlgorithms/attGuidance/_GeneralModuleFiles/constrainedAxisPointingLibrary.h"
 %include "architecture/msgPayloadDefC/NavAttMsgPayload.h"
 struct NavAttMsg_C;
 %include "architecture/msgPayloadDefC/BodyHeadingMsgPayload.h"
